Abstracting data holder
