+++ frontend/src/Features/ManageEntity/Author/ManageAuthorViewModel.tsx
@@ -1,71 +1,22 @@
 import { AxiosResponse } from 'axios';
 import {
-  map, sort, slice, range, reverse, filter, compose, curry,
+  map, range,
 } from 'ramda';
-import { useEffect, useMemo, useState } from 'react';
+import { useEffect, useMemo } from 'react';
 import AuthorAPI from '../../../API/AuthorAPI';
 import AuthorSorter from '../../../Functions/Helpers/AuthorSorter';
-import IntBiFunction from '../../../Functions/Interfaces/IntBiFunction';
+import Predicate from '../../../Functions/Interfaces/Predicate';
 import EntityMapper from '../../../Functions/Mappers/EntityMapper';
 import ManageAuthorMapper from '../../../Functions/Mappers/ManageAuthorMapper';
 import AuthorEntity from '../../../Types/Entities/AuthorEntity';
 import SortOrder from '../../../Types/SortOrder';
-
-function useAuthorsDataHolder() {
-  const [authorsData, setAuthorsData] = useState<AuthorEntity[]>([]);
-  const [preparedData, setPreparedData] = useState<AuthorEntity[]>([]);
-
-  const [query, setQuery] = useState('');
-  const [sortOrder, setSortOrder] = useState<SortOrder>('asc');
-  const [sortBy, setSortBy] = useState<IntBiFunction<AuthorEntity>>(() => AuthorSorter.byName);
-  const [showsPerPage, setShowsPerPage] = useState(10);
-  const [page, setPage] = useState(1);
-
-  const totalPage = (authorsData.length / showsPerPage) + 1;
-
-  const setData = (entities: AuthorEntity[]) => {
-    setAuthorsData(entities);
-    setPreparedData(entities);
-  };
-
-  const filterByQuery = (value: AuthorEntity) => value.name.toLowerCase().includes(
-    query.toLowerCase(),
-  );
-
-  const sortOrderFunc = (data: AuthorEntity[]) => (sortOrder === 'asc' ? data : reverse(data));
-
-  useEffect(() => {
-    if (authorsData.length === 0) return;
-    const offset = (showsPerPage * page) - showsPerPage;
-    const sliceByOffset = slice(offset, offset + showsPerPage);
-
-    const filterPipeline = compose(
-      sortOrderFunc,
-      sort(sortBy),
-      filter(filterByQuery),
-    );
-
-    let data = filterPipeline(authorsData);
-    data = sliceByOffset(data);
-    setPreparedData(data);
-  }, [authorsData, query, page, showsPerPage, sortBy, sortOrder]);
-
-  return {
-    authorsData: preparedData,
-    setData,
-    setQuery,
-    setSortBy,
-    setSortOrder,
-    showsPerPage,
-    setShowsPerPage,
-    page,
-    setPage,
-    totalPage,
-  };
-}
+import useEntityDataHolder, { EntityFilterByQuery } from '../EntityDataHolder';
 
 export default function useManageAuthorViewModel() {
-  const dataHolder = useAuthorsDataHolder();
+  const filterByQuery: EntityFilterByQuery<AuthorEntity> = (query: string):
+  Predicate<AuthorEntity> => (value) => value.name.toLowerCase().includes(query.toLowerCase());
+
+  const dataHolder = useEntityDataHolder(AuthorSorter.byName, filterByQuery);
 
   const onFetchSuccess = (response: AxiosResponse) => {
     const data = map(EntityMapper.author, response.data);
@@ -78,11 +29,10 @@ export default function useManageAuthorViewModel() {
 
   useEffect(() => { fetchAuthors(); }, []);
 
-  const startIndex = (dataHolder.page * dataHolder.showsPerPage) - dataHolder.showsPerPage;
-  const mapper = ManageAuthorMapper.curriedTableRow(startIndex);
+  const mapper = ManageAuthorMapper.curriedTableRow(dataHolder.startIndex);
   const authorsElement = useMemo(
-    () => dataHolder.authorsData.map(mapper),
-    [dataHolder.authorsData],
+    () => dataHolder.entityData.map(mapper),
+    [dataHolder.entityData],
   );
 
   const sortByElement = useMemo(() => (
