+++ frontend/src/Functions/Interfaces/Predicate.ts
@@ -1,0 +1,3 @@
+type Predicate<T> = (value: T) => boolean;
+
+export default Predicate;
